Make install instructions match git directory
Currently the custom component in the git repo is in `warmup_cc`, so I believe the cp in the install instructions should copy the contents from the `warmup_cc` putting them in `<configpath>/custom_components/warmup`. 

(It's was a little confusing because at first it looks like the warmup4ie-PyPi and warmup_cc might need to be copied, but there is a (currently) identical copy of the warmup4ie module under `warmup_cc` so the `warmup4ie-PyPi` directory can be ignored for now.